# apps/api/openapi.yaml
openapi: 3.0.0
info:
  title: Project Management System API
  version: 1.0.0
  description: REST API for the Project Management System
servers:
  - url: https://pms-api-production-bad8.up.railway.app
    description: Production Server
  - url: http://localhost:3000
    description: Local Server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fullName:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [NOT_STARTED, IN_PROGRESS, COMPLETED]
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, COMPLETED]
        dueDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            details:
              type: array
              items:
                type: string

security:
  - bearerAuth: []

paths:
  /api/auth/register:
    post:
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, email, password]
              properties:
                fullName:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/login:
    post:
      summary: Authenticate user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/logout:
    post:
      summary: Logout user
      responses:
        '200':
          description: Logged out successfully

  /api/auth/me:
    get:
      summary: Get current authenticated user
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized

  /api/projects:
    get:
      summary: List user projects
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Search term for project name or description
        - in: query
          name: status
          schema:
            type: string
            enum: [NOT_STARTED, IN_PROGRESS, COMPLETED]
          description: Filter by status
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
    post:
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, startDate, endDate]
              properties:
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [NOT_STARTED, IN_PROGRESS, COMPLETED]
                startDate:
                  type: string
                  format: date-time
                endDate:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Project'

  /api/projects/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get project details
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '404':
          description: Project not found
    put:
      summary: Update project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [NOT_STARTED, IN_PROGRESS, COMPLETED]
                startDate:
                  type: string
                  format: date-time
                endDate:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Project'
        '404':
          description: Project not found
    delete:
      summary: Delete project
      responses:
        '204':
          description: Project deleted
        '404':
          description: Project not found

  /api/tasks:
    get:
      summary: List tasks
      parameters:
        - in: query
          name: projectId
          schema:
            type: string
            format: uuid
          description: Filter tasks by project ID
        - in: query
          name: search
          schema:
            type: string
          description: Search term for task name or description
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, IN_PROGRESS, COMPLETED]
          description: Filter by status
        - in: query
          name: priority
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH]
          description: Filter by priority
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
    post:
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, name, dueDate]
              properties:
                projectId:
                  type: string
                  format: uuid
                name:
                  type: string
                description:
                  type: string
                priority:
                  type: string
                  enum: [LOW, MEDIUM, HIGH]
                status:
                  type: string
                  enum: [PENDING, IN_PROGRESS, COMPLETED]
                dueDate:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '404':
          description: Project not found

  /api/tasks/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get task details
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
    put:
      summary: Update task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                priority:
                  type: string
                  enum: [LOW, MEDIUM, HIGH]
                status:
                  type: string
                  enum: [PENDING, IN_PROGRESS, COMPLETED]
                dueDate:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
    delete:
      summary: Delete task
      responses:
        '204':
          description: Task deleted
        '404':
          description: Task not found

  /api/dashboard:
    get:
      summary: Get user dashboard metrics
      responses:
        '200':
          description: Dashboard metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      totalProjects:
                        type: integer
                      activeProjects:
                        type: integer
                      totalTasks:
                        type: integer
                      completedTasks:
                        type: integer
                      pendingTasks:
                        type: integer